USSF should be opensource
I never knew this little tool became popular after I
stopped visiting MSFN [1] on a regular basis.

TBH, it's not nice code, but it's now open :)

NOTE: When archiving this as RAR or other familiar formats,
some anti-virus software reported false positives, probably
because of the "similar" binary unpacking tools in bin/.

If you are afraid this is a real threat, ignore the "/bin"
directory and its tools, which you can fetch from trusted
sources instead.

[1] http://www.msfn.org/board/

Signed-off-by: Alexandru Avadanii <alex@softwareelite.com>